Moved model choices from vocab
